+++ cfg/config.php
@@ -3,7 +3,7 @@
 define('JOM_DEBUG',     true);
 define('JOM_LOG',       true);
 define('JOM_MILESTONE', '0.3');                                         // TICKET CREATION version (refer to MILESTONES.md for versioning progression)
-define('JOM_RELEASE',   '1');
+define('JOM_RELEASE',   '2');
 define('JOM_VERSION',   JOM_MILESTONE.'.'.JOM_RELEASE);                 // milestone.release
 define('JOM_DESC_VER',  'v. '.JOM_MILESTONE.'.x: ticket creation');     // milestone version description
 
@@ -33,6 +33,7 @@ define('TBL_COMPANIES',           TABLES_PREFIX.'Company');
 define('TBL_USERS',               TABLES_PREFIX.'Users');
 define('TBL_JOBS',                TABLES_PREFIX.'Jobs');
 define('TBL_USERS_ACL',           TABLES_PREFIX.'Users_acl');
+define('TBL_CATEGORIES_GENERIC',  TABLES_PREFIX.'Categories');
 define('TBL_CATEGORIES_1',        TABLES_PREFIX.'Categories_1');
 define('TBL_CATEGORIES_2',        TABLES_PREFIX.'Categories_2');
 define('TBL_LOGIN_ATTEMPTS',      TABLES_PREFIX.'Login_attempts');